re-architect script to download and play faster...
...and add/clarify some command-line options. Adds support for updating
metadata for a single word. Adds support for forcing files to download
for a word. Fixes issues with regexes (though still brittle). Populates
more metadata.

New architecture is more resilient to the situation when a file failed
to download. A threadpool is used to download the remaining files after
the first file is played.